Add files via upload
The files provided are an implementation of a doubly linked list that we can use to model the tables. Please note that the linked list has not been genericized yet (that is, I have not added any template functionality to it). The code has been uploaded to allow collaborators to test the linked list to test the linked list further. There is a double free error that has not been resolved. If possible, attempt to resolve the matter and push the code back. In the event that the matter is resolved before any commits, the code will be committed to the development branch.