State Order ModifiedZ
I have added comments to functions and variables and modified all classes related to the state pattern. What I need is for Future to access the cook method for each state and decide what the order should do in each state, or you can decide not to do anything?
I have also modified the main, please look at it to see if it works well with the ChefHandler.